+++ cypress/support/e2e.ts
@@ -16,7 +16,7 @@
 // Import commands.js using ES2015 syntax:
 import './commands'
 import GravityCollector from '../../src'
-import { Listener } from '../../src/types'
+import { CookieStrategy, Listener } from '../../src/types'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
@@ -28,5 +28,6 @@ Cypress.on('window:load', (win) => {
     gravityServerUrl: 'https://api.gravity.smartesting.com',
     window: win,
     enabledListeners: [Listener.Click, Listener.KeyUp, Listener.KeyDown, Listener.Change],
+    cookieStrategy: CookieStrategy.subDomains,
   })
 })