+++ src/test-utils/clearCookies.ts
@@ -1,0 +1,5 @@
+export function clearCookies() {
+  document.cookie.split(';').forEach(function (c) {
+    document.cookie = c.replace(/^ +/, '').replace(/=.*/, `=;expires=${new Date().toUTCString()};path=/`)
+  })
+}